Refactor the process_transaction function
I decomposed this into several sub-functions. It'll be easier to break
this out into modules later on, if I choose to. But it makes the
overall intent of the match statement a lot clearer.

I still need to think about if I want to factor out some of the common
logic in these functions, but that's going to be tough given how
particular each one is. They all have subtle differences, so factoring
out common parts is tough.