Set the amount type from f32 to f64
So it turns out that after ~1m transactions, f32 loses a lot of
precision. To the extent that if you run 1m deposits of 50.0 each,
you'd lose over 500k of credit. Which is not cool. So spend the RAM
for better precision.

The _smart_ thing to do here would be to do something like a decicents
type, something where the value is stored internally as an integer,
but is always rendered in terms of dollars. Getting that to play
nicely with serde and csv_async would be a ... challenge.