Add a test for 1m rows.
I ran into something fun when testing manually, which is that f32 will
lose a lot of precision if you run 1m deposits of 50.00 each. To the
tune of the client losing over 500k. Switching to f64 fixes that, but
it still rankles me. I may see if there's a decimal type or crate
lying around that plays nicely with serde.