Add comments, tests, and change the default filename arg.
I added a bunch of comments for the types and functions in this
project, to explain the why behind the how.

I also decided that you cannot resolve a transaction for a locked
account. I think that point is where a human should step in and make a
judgment call, rather than relying on an automated process.

I added tests for this, and changed how the program handles a missing
file argument. Previously, I'd had this default to a file I knew was
around, so that running it from an emacs hotkey was much easier. But
for the shipping/"release" version, I want to give the user an error
rather than just randomly reading a file they perhaps don't know
about.

I also deleted the "LedgerEntryEnum" struct. This was an attempt by me
to combine the TransactionType sum type with the LedgerEntry struct,
and avoid having an Optional on Amount. Serde didn't quite get what I
was trying to do, alas.