Break code out into modules!
Time to stop putting this off. It was actually fairly painless, the
only "test-induced damage" was marking process_transaction (singular)
public when it isn't strictly necessary to do so.

I suppose I could have moved the tests for it into the module, but
there are a _lot_ of them, and I like having the files be small and
easier to move around in.